Remove BaseSsh* classes
Fixes #18. These were leftover from an abstraction layer to support an
alternative SSH implementation that use a slightly different protocol.
That implementation was never part of Giraffe and is no longer
supported, so these classes just add complexity.
