+++ ssh/src/main/java/com/palantir/giraffe/ssh/SshHost.java
@@ -17,12 +17,21 @@ package com.palantir.giraffe.ssh;
 
 import java.io.IOException;
 import java.net.URI;
+import java.nio.file.FileSystem;
+import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Map;
 
+import com.palantir.giraffe.SystemConverter;
+import com.palantir.giraffe.command.ExecutionSystem;
+import com.palantir.giraffe.file.base.SuppressedCloseable;
+import com.palantir.giraffe.host.AbstractHostControlSystem;
 import com.palantir.giraffe.host.Host;
+import com.palantir.giraffe.host.HostControlSystem;
+import com.palantir.giraffe.host.RemoteHostAccessor;
+import com.palantir.giraffe.ssh.internal.SshEnvironments;
 import com.palantir.giraffe.ssh.internal.SshUris;
-import com.palantir.giraffe.ssh.internal.base.BaseSshHostAccessor;
 
 /**
  * Provides access to a file system and execution system on a host using SSH.
@@ -31,7 +40,7 @@ import com.palantir.giraffe.ssh.internal.base.BaseSshHostAccessor;
  *
  * @param <C> the type of credential used to access the host
  */
-public final class SshHost<C extends AbstractSshCredential> extends BaseSshHostAccessor<C> {
+public final class SshHost<C extends AbstractSshCredential> implements RemoteHostAccessor<C> {
 
     private static final int DEFAULT_SSH_PORT = 22;
 
@@ -84,13 +93,57 @@ public final class SshHost<C extends AbstractSshCredential> extends BaseSshHostA
         return new SshHost<C>(host, port, credential);
     }
 
+    private final Host host;
+    private final int port;
+    private final C credential;
+
     private SshHost(Host host, int port, C credential) {
-        super(host, port, credential);
+        this.host = host;
+        this.port = port;
+        this.credential = credential;
     }
 
-    @Override
     public URI getSystemUri() {
         return SshUris.getUri(getHost(), getPort(), getCredential());
     }
 
+    @Override
+    public Host getHost() {
+        return host;
+    }
+
+    public int getPort() {
+        return port;
+    }
+
+    @Override
+    public C getCredential() {
+        return credential;
+    }
+
+    public Map<String, ?> getSystemEnvironment() {
+        return SshEnvironments.makeEnv(this);
+    }
+
+    @Override
+    public HostControlSystem openHostControlSystem() throws IOException {
+        FileSystem fs = FileSystems.newFileSystem(
+                getSystemUri(),
+                getSystemEnvironment(),
+                getClass().getClassLoader());
+
+        // use the file system as the canonical system source
+        return new SshHostControlSystem(host, fs, SystemConverter.asExecutionSystem(fs));
+    }
+
+    private static final class SshHostControlSystem extends AbstractHostControlSystem {
+        private SshHostControlSystem(Host host, FileSystem fs, ExecutionSystem es) {
+            super(host, fs, es);
+        }
+
+        @Override
+        public void close() throws IOException {
+            SuppressedCloseable.create(getExecutionSystem(), getFileSystem()).close();
+        }
+    }
 }
