+++ ssh/src/main/java/com/palantir/giraffe/ssh/internal/SshPath.java
@@ -15,16 +15,28 @@
  */
 package com.palantir.giraffe.ssh.internal;
 
+import java.io.IOException;
+import java.nio.file.LinkOption;
 import java.util.Objects;
 
+import javax.annotation.CheckForNull;
+
+import com.palantir.giraffe.command.CommandResult;
+import com.palantir.giraffe.file.base.AbstractImmutableListPath;
 import com.palantir.giraffe.file.base.ImmutableListPathCore;
-import com.palantir.giraffe.ssh.internal.base.BaseSshFileSystem;
-import com.palantir.giraffe.ssh.internal.base.BaseSshPath;
 
-final class SshPath extends BaseSshPath<SshPath> {
+import net.schmizz.sshj.sftp.SFTPClient;
 
-    SshPath(BaseSshFileSystem<SshPath> fs, ImmutableListPathCore core) {
-        super(fs, core);
+final class SshPath extends AbstractImmutableListPath<SshPath> {
+
+    protected static final ImmutableListPathCore.Parser PARSER =
+            ImmutableListPathCore.parser(SshFileSystem.SEPARATOR);
+
+    private final SshFileSystem fs;
+
+    SshPath(SshFileSystem fs, ImmutableListPathCore core) {
+        super(core);
+        this.fs = fs;
     }
 
     @Override
@@ -47,6 +59,37 @@ final class SshPath extends BaseSshPath<SshPath> {
             SshPath that = (SshPath) obj;
             return getFileSystem().equals(that.getFileSystem())
                 && toString().equals(that.toString());
+        }
+    }
+
+    @Override
+    public SshFileSystem getFileSystem() {
+        return fs;
+    }
+
+    @CheckForNull
+    public String getInode() throws IOException {
+        CommandResult result = getFileSystem().execute("stat", "-c", "%i", this);
+        if (result.getExitStatus() == 0) {
+            return result.getStdOut().trim();
+        } else {
+            return null;
+        }
+    }
+
+    @Override
+    public SshPath toRealPath(LinkOption... options) throws IOException {
+        for (LinkOption option : options) {
+            if (option == LinkOption.NOFOLLOW_LINKS) {
+                throw new UnsupportedOperationException("NOFOLLOW_LINKS not supported");
+            } else {
+                throw new IllegalArgumentException("unknown option: " + option);
+            }
+        }
+
+        try (SFTPClient sftp = getFileSystem().openSftpClient()) {
+            String realPath = sftp.canonicalize(toString());
+            return newPath(PARSER.parse(realPath));
         }
     }
 }
