+++ ssh/src/main/java/com/palantir/giraffe/ssh/internal/SshSystemContext.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.giraffe.ssh.internal.base;
+package com.palantir.giraffe.ssh.internal;
 
 import java.net.URI;
 
@@ -24,7 +24,7 @@ import org.slf4j.Logger;
  *
  * @author bkeyes
  */
-public class SshSystemContext {
+class SshSystemContext {
 
     private final URI uri;
     private final SharedSshClient client;
